requirements.txt
# Core application and UI
streamlit
python-dotenv

# LangChain and LangGraph for agentic workflow
langchain
langchain-core
langgraph

# Google Generative AI for LLM access
google-generativeai

# Web scraping and browser automation
selenium
undetected-chromedriver

# Data handling and Excel parsing
pandas
openpyxl

# PDF generation and text extraction
reportlab
PyMuPDF
pypdf

matplotlib
tabulate

# Async support for optimized web operations
aiohttp
asyncio

# for colored terminal output
colorama
lxml

requests
beautifulSoup4

tavily-python
nest_asyncio

google-genai 
pyaudio 
asyncio 
termcolor
websockets

